moving Header to an independent reuabsle component state
