+++ imports/ui/LayoutContainer.jsx
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 
 import {
@@ -9,27 +9,17 @@ import {
   ChakraProvider,
   Flex,
   Heading,
-  HStack,
   Image,
   Link as CLink,
   List,
   ListItem,
-  Menu as CMenu,
-  MenuButton,
-  MenuList,
-  MenuItem,
   Spinner,
   Text,
-  Wrap,
 } from '@chakra-ui/react';
-import { ChevronDownIcon } from '@chakra-ui/icons';
-import { Container, Row, Col, ScreenClassRender } from 'react-grid-system';
 import { Helmet } from 'react-helmet';
-
 import { useTranslation } from 'react-i18next';
 
 import Hosts from '../api/hosts/host';
-import UserPopup from './components/UserPopup';
 import ChangeLanguage from './components/ChangeLanguageMenu';
 import FeedbackForm from './components/FeedbackForm';
 import { chakraTheme } from './utils/constants/theme';
@@ -37,45 +27,6 @@ import { chakraTheme } from './utils/constants/theme';
 export const StateContext = React.createContext(null);
 
 const publicSettings = Meteor.settings.public;
-
-const getRoute = (item, index) => {
-  if (index === 0) {
-    return '/';
-  }
-  if (item.name === 'info') {
-    return '/pages/about';
-  }
-  return `/${item.name}`;
-};
-
-const getBackgroundStyle = (cHue) => {
-  return {
-    backgroundColor: 'rgb(255, 255, 255)',
-  };
-
-  if (!cHue) {
-    return {
-      backgroundColor: '#fff',
-    };
-  }
-  return {
-    backgroundImage: `-moz-linear-gradient(
-    0deg,
-    hsl(${cHue}, 80%, 80%),
-    hsl(${cHue}, 80%, 95%)
-  )`,
-    backgroundImage: `-webkit-linear-gradient(
-    0deg,
-    hsl(${cHue}, 80%, 80%),
-    hsl(${cHue}, 80%, 95%)
-  )`,
-    backgroundImage: `linear-gradient(
-    0deg,
-    hsl(${cHue}, 80%, 99%),
-    hsl(${cHue}, 80%, 90%)
-  )`,
-  };
-};
 
 function LayoutPage({ currentUser, currentHost, userLoading, hostLoading, history, children }) {
   const [tc] = useTranslation('common');
@@ -107,13 +58,6 @@ function LayoutPage({ currentUser, currentHost, userLoading, hostLoading, histor
     );
   }
 
-  const headerProps = {
-    currentUser,
-    currentHost,
-    history,
-    title: 'Fanus',
-  };
-
   const hostWithinUser =
     currentUser &&
     currentUser.memberships &&
@@ -121,9 +65,6 @@ function LayoutPage({ currentUser, currentHost, userLoading, hostLoading, histor
 
   const role = hostWithinUser && hostWithinUser.role;
   const canCreateContent = role && ['admin', 'contributor'].includes(role);
-  const { menu } = currentHost.settings;
-  // console.log(pathname)
-  const isHeader = menu.some((item) => item.path === pathname);
 
   return (
     <ChakraProvider theme={chakraTheme}>
@@ -141,10 +82,8 @@ function LayoutPage({ currentUser, currentHost, userLoading, hostLoading, histor
           canCreateContent,
         }}
       >
-        <Center className="main-viewport" style={getBackgroundStyle()}>
+        <Center className="main-viewport">
           <Box maxWidth="1400px" w="100%">
-            {<Header {...headerProps} />}
-
             <Box style={{ minHeight: '100vh' }}>{children}</Box>
 
             <Footer currentHost={currentHost} tc={tc} />
@@ -152,131 +91,6 @@ function LayoutPage({ currentUser, currentHost, userLoading, hostLoading, histor
         </Center>
       </StateContext.Provider>
     </ChakraProvider>
-  );
-}
-
-function Header({ currentUser, currentHost, title, history }) {
-  return (
-    <Box mb="4">
-      <ScreenClassRender
-        render={(screenClass) => {
-          const isMobile = ['xs'].includes(screenClass);
-
-          return (
-            <Container fluid style={{ width: '100%', padding: 0, zIndex: 9 }}>
-              <Row
-                style={{
-                  marginLeft: 0,
-                  marginRight: 0,
-                  marginBottom: 12,
-                  marginTop: 12,
-                  alignItems: 'flex-start',
-                }}
-              >
-                <Col xs={3} style={{ paddingLeft: 0 }}>
-                  <Link to="/">
-                    <Box w="120px" h="60px" ml="3">
-                      <Image
-                        fit="contain"
-                        src={currentHost && currentHost.logo}
-                        className="header-logo"
-                      />
-                    </Box>
-                  </Link>
-                </Col>
-                <Col xs={6} style={{ display: 'flex', justifyContent: 'center' }}>
-                  {!isMobile && (
-                    <Menu
-                      currentHost={currentHost}
-                      isMobile={false}
-                      screenClass={screenClass}
-                      history={history}
-                    />
-                  )}
-                </Col>
-                <Col xs={3} style={{ paddingRight: 0 }}>
-                  <Flex justify="flex-end">
-                    <UserPopup currentUser={currentUser} />
-                  </Flex>
-                </Col>
-              </Row>
-              {isMobile && <Menu currentHost={currentHost} isMobile history={history} />}
-            </Container>
-          );
-        }}
-      />
-    </Box>
-  );
-}
-
-function Menu({ currentHost, isMobile, screenClass, history }) {
-  const ref = useRef();
-  if (!currentHost || !currentHost.settings || !currentHost.settings.menu) {
-    return null;
-  }
-
-  const menu = currentHost.settings.menu;
-
-  const menuItems = menu
-    .filter((item) => item.isVisible)
-    .map((item, index) => ({
-      ...item,
-      route: getRoute(item, index),
-    }));
-
-  const pathname = history.location.pathname;
-
-  const handleClick = (item) => {
-    history.push(item.route);
-  };
-
-  const isCurrentPage = (name) => {
-    if (name === 'info') {
-      return pathname.substring(0, 6) === '/pages';
-    }
-    return name === pathname.substring(1, pathname.length);
-  };
-
-  const activeMenuItem = menuItems.find((item) => isCurrentPage(item.name));
-
-  if (['lg', 'xl', 'xxl'].includes(screenClass)) {
-    return (
-      <Wrap align="center" pt="lg" spacing="4">
-        {menuItems.map((item) => (
-          <Box as="button" key={item.name} onClick={() => handleClick(item)}>
-            <Text
-              borderBottom={
-                activeMenuItem && activeMenuItem.label === item.label ? '1px solid #010101' : 'none'
-              }
-              mx="1"
-              textTransform="capitalize"
-            >
-              {item.label}
-            </Text>
-          </Box>
-        ))}
-      </Wrap>
-    );
-  }
-
-  return (
-    <Box align="center">
-      <CMenu placement="bottom" closeOnSelect>
-        <MenuButton>
-          <HStack>
-            <Text textTransform="capitalize">{activeMenuItem ? activeMenuItem.label : 'Menu'}</Text>
-            <ChevronDownIcon />
-          </HStack>
-        </MenuButton>
-        <MenuList>
-          {menuItems.map((item) => (
-            <MenuItem key={item.label} onClick={() => handleClick(item)}>
-              {item.label}
-            </MenuItem>
-          ))}
-        </MenuList>
-      </CMenu>
-    </Box>
   );
 }
