+++ imports/ui/pages/activities/Activities.jsx
@@ -11,6 +11,7 @@ import { StateContext } from '../../LayoutContainer';
 import Loader from '../../components/Loader';
 import Paginate from '../../components/Paginate';
 import PublicActivityThumb from '../../components/PublicActivityThumb';
+import Header from '../../components/Header';
 
 moment.locale(i18n.language);
 
@@ -66,6 +67,8 @@ function Activities({ activitiesList, isLoading, history }) {
           currentHost.settings.name
         } | ${publicSettings.name}`}</title>
       </Helmet>
+
+      <Header />
 
       <Center mb="4">
         {canCreateContent && (
